CREATIS Informatique

+++ app/src/main/java/ch/creatis/bexio/First/4. TempsActivity.kt
@@ -137,7 +137,7 @@ class TempsActivity : AppCompatActivity() {
 
 
 
-                        val idBexio= response.getJSONObject(i)["id"].toString()
+                        val idBexio= response.getJSONObject(i)["id"].toString().toInt()
 
                         var userId = response.getJSONObject(i)["user_id"].toString().toInt()
