+++ app/src/main/java/ch/creatis/bexio/Login/0. LoginActivity.kt
@@ -873,7 +873,7 @@ import kotlin.collections.HashMap
 
 
 
-                val idBexio= response.getJSONObject(i)["id"].toString()
+                val idBexio= response.getJSONObject(i)["id"].toString().toInt()
 
                 var userId = response.getJSONObject(i)["user_id"].toString().toInt()
